remove dfs routes
