+++ ingest/sqs/router/usecase/export_test.go
@@ -19,10 +19,6 @@ const (
 	NoTotalValueLockedError = noTotalValueLockedError
 )
 
-func (r Router) FindRoutes(tokenInDenom, tokenOutDenom string, currentRoute domain.Route, poolsUsed []bool, previousTokenOutDenoms []string) ([]domain.Route, error) {
-	return r.findRoutes(tokenInDenom, tokenOutDenom, currentRoute, poolsUsed, previousTokenOutDenoms)
-}
-
 func (r Router) GetBestSplitRoutesQuote(routes []domain.Route, tokenIn sdk.Coin) (quote domain.Quote, err error) {
 	return r.estimateBestSplitRouteQuote(routes, tokenIn)
 }
@@ -41,4 +37,15 @@ func (r *routerUseCaseImpl) HandleRoutes(ctx context.Context, router *Router, to
 
 func (r *Router) GetOptimalQuote(tokenIn sdk.Coin, tokenOutDenom string, routes []domain.Route) (domain.Quote, error) {
 	return r.getOptimalQuote(tokenIn, tokenOutDenom, routes)
+}
+
+// GetSortedPoolIDs returns the sorted pool IDs.
+// The sorting is initialized in NewRouter() by preferredPoolIDs and TVL.
+// Only used for tests.
+func (r Router) GetSortedPoolIDs() []uint64 {
+	sortedPoolIDs := make([]uint64, len(r.sortedPools))
+	for i, pool := range r.sortedPools {
+		sortedPoolIDs[i] = pool.GetId()
+	}
+	return sortedPoolIDs
 }
